carbontrack/green_project/requirements.txt
django
djangorestframework
django-cors-headers
gunicorn
psycopg2-binary
whitenoise
python-decouple